remove the optimizer base and learning rate base (#56099)
* remove the optimizer base and learning rate base

* fix bug

* fix bug
