+++ python/paddle/fluid/dygraph/__init__.py
@@ -18,9 +18,6 @@ from .base import *
 from . import tracer
 from .tracer import *
 
-from . import learning_rate_scheduler
-from .learning_rate_scheduler import *
 
 __all__ = []
 __all__ += base.__all__
-__all__ += learning_rate_scheduler.__all__
