+++ python/paddle/optimizer/optimizer.py
@@ -1114,9 +1114,6 @@ class Optimizer:
         end = len(target_block.ops)
         return target_block._slice_ops(start, end)
 
-    def _append_dgc_ops(self, param_and_grad):
-        pass
-
     def backward(
         self,
         loss,
@@ -1205,9 +1202,6 @@ class Optimizer:
                     params_grads = append_backward(
                         loss, parameter_list, act_no_grad_set, callbacks
                     )
-                # Note: since we can't use all_reduce_op now,
-                #  dgc_op should be the last op of one grad.
-                self._append_dgc_ops(params_grads)
         return params_grads
 
     def apply_gradients(self, params_grads):
