+++ python/paddle/static/__init__.py
@@ -70,7 +70,6 @@ from ..fluid.framework import ipu_shard_guard  # noqa: F401
 from ..fluid.framework import set_ipu_shard  # noqa: F401
 from .nn.control_flow import Print  # noqa: F401
 from ..fluid.param_attr import WeightNormParamAttr  # noqa: F401
-from ..fluid.optimizer import Optimizer  # noqa: F401
 
 
 from .nn.metric import auc  # noqa: F401
